🩹 [Patch]: Add macOS and Windows tests (#5)
## Description

This pull request includes changes to the
`.github/workflows/Action-Test.yml` file to enhance the testing strategy
by incorporating a matrix for operating systems.

Improvements to testing strategy:

*
[`.github/workflows/Action-Test.yml`](diffhunk://#diff-a12ae5c885b0673c0ff6f70c2670886907590d624626e07da4c52e01aeaf56a4R21-R25):
Added a matrix strategy to test on multiple operating systems
(`ubuntu-latest`, `windows-latest`, `macos-latest`) and updated the
`runs-on` field to use the matrix variable.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas